+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api/AnonymousApi.kt
@@ -1,0 +1,11 @@
+package io.getstream.chat.android.client.api
+
+/**
+ * Marks an API as requiring no authentication.
+ *
+ * For differences between authenticated and anonymous behaviour, see
+ * [HeadersInterceptor] and [TokenAuthInterceptor].
+ */
+@Retention(AnnotationRetention.RUNTIME)
+@Target(AnnotationTarget.CLASS)
+internal annotation class AnonymousApi