+++ stream-chat-android-sample/src/main/kotlin/io/getstream/chat/sample/feature/login/LoginModule.kt
@@ -1,8 +1,0 @@
-package io.getstream.chat.sample.feature.login
-
-import org.koin.androidx.viewmodel.dsl.viewModel
-import org.koin.dsl.module
-
-val loginModule = module {
-    viewModel { LoginViewModel(get(), get()) }
-}