+++ stream-chat-android/src/main/java/com/getstream/sdk/chat/view/MessageListView.kt
@@ -154,7 +154,7 @@ public class MessageListView : ConstraintLayout {
         }
     private val DEFAULT_ATTACHMENT_CLICK_LISTENER =
         AttachmentClickListener { message, attachment ->
-            Chat.getInstance()
+            Chat.instance()
                 .navigator
                 .navigate(AttachmentDestination(message, attachment, context))
         }